# backend/internal/auth/infrastructure/controllers/logout-controller.go
package controllers

import (
	"net/http"

	api "rodrigoorlandini/vet-shifter/internal/_shared/api"
	authmiddleware "rodrigoorlandini/vet-shifter/internal/auth/infrastructure/middleware"

	"github.com/gin-gonic/gin"
)

type LogoutController struct{}

func NewLogoutController() *LogoutController {
	return &LogoutController{}
}

// Logout godoc
//
//	@Summary		Logout (client should discard token)
//	@Description	Stateless JWT: server does not invalidate. Client must clear stored token.
//	@Tags			auth
//	@Security		BearerAuth
//	@Produce		json
//	@Success		200	"OK"
//	@Failure		401	{object}	api.ApiErrorResponse	"Não autorizado"
//	@Router			/auth/logout [post]
func (c *LogoutController) Handle(ctx *gin.Context) {
	authmiddleware.ClearAccessTokenCookie(ctx)
	ctx.JSON(http.StatusOK, gin.H{"message": "Sessão encerrada."})
}

var _ = api.ApiErrorResponse{}